+++ classes/ui/wpdk-menu.php
@@ -121,16 +121,39 @@ class WPDKMenu {
   }
 
   /**
+   * Return a sanitize view controller for a callable
+   *
+   * @brief Sanitize
+   *
+   * @param string|array $view_controller Callable
+   *
+   * @return string
+   */
+  public static function sanitizeViewController( $view_controller )
+  {
+    if ( is_string( $view_controller ) ) {
+      $result = $view_controller;
+    }
+    elseif ( is_array( $view_controller ) ) {
+      $result = get_class( $view_controller[0] ) . '-' . $view_controller[1];
+    }
+    return $result;
+  }
+
+  /**
    * Return the WPDK menu info by name of view controller of submenu item
    *
-   * @param string $view_controller The view controller class name
+   * @param string|array $view_controller Any callable or view controller class name
    *
    * @return array
    */
-  public static function menu( $view_controller ) {
+  public static function menu( $view_controller )
+  {
+    $global_key = self::sanitizeViewController( $view_controller );
+
     if ( isset( $GLOBALS[self::GLOBAL_MENU] ) ) {
-      if ( !empty( $view_controller ) && !empty( $GLOBALS[self::GLOBAL_MENU][$view_controller] ) ) {
-        return $GLOBALS[self::GLOBAL_MENU][$view_controller];
+      if ( !empty( $global_key ) && !empty( $GLOBALS[self::GLOBAL_MENU][$global_key] ) ) {
+        return $GLOBALS[self::GLOBAL_MENU][$global_key];
       }
     }
     else {
@@ -627,24 +650,28 @@ class WPDKSubMenu {
 
     global $plugin_page;
 
-    $hook = '';
+    $hook       = '';
+    $global_key = WPDKMenu::sanitizeViewController( $this->viewController );
 
     if ( !empty( $this->viewController ) ) {
-      if( is_string( $this->viewController ) && !function_exists( $this->viewController ) ) {
+      if ( is_string( $this->viewController ) && !function_exists( $this->viewController ) ) {
+        /* @todo Think $vc = %s::init() - in this way we can use the singleton in the head hook below */
         $hook = create_function( '', sprintf( '$view_controller = new %s; $view_controller->display();', $this->viewController ) );
-
-        /* Create a global list of my own menu. */
-        $GLOBALS[ WPDKMenu::GLOBAL_MENU ][$this->viewController] = array(
-          'parent'     => $this->parent,
-          'page'       => $this->id,
-          'hook'       => '',
-          'menu_title' => ''
-        );
       }
       // If the callable is in the form array( obj, method ), I have to properly init $hook anyway
-      elseif( is_callable( $this->viewController ) ) {
+      elseif ( is_callable( $this->viewController ) ) {
         $hook = $this->viewController;
       }
+    }
+
+    if( !empty( $global_key ) ) {
+      /* Create a global list of my own menu. */
+      $GLOBALS[WPDKMenu::GLOBAL_MENU][$global_key ] = array(
+        'parent'     => $this->parent,
+        'page'       => $this->id,
+        'hook'       => '',
+        'menu_title' => ''
+      );
     }
 
     /* Apply filter for change the title. */
@@ -671,15 +698,16 @@ class WPDKSubMenu {
     }
 
     if ( !empty( $this->viewController ) && is_string( $this->viewController ) && !function_exists( $this->viewController ) ) {
-
-      $GLOBALS[WPDKMenu::GLOBAL_MENU][$this->viewController]['hook']       = $this->hookName;
-      $GLOBALS[WPDKMenu::GLOBAL_MENU][$this->viewController]['menu_title'] = $menu_title;
-
       $will_load = create_function( '', sprintf( '%s::willLoad();', $this->viewController ) );
       add_action( 'load-' . $this->hookName, $will_load );
 
       $head = create_function( '', sprintf( '%s::didHeadLoad();', $this->viewController ) );
       add_action( 'admin_head-' . $this->hookName, $head );
+    }
+
+    if ( !empty( $global_key ) ) {
+      $GLOBALS[WPDKMenu::GLOBAL_MENU][$global_key]['hook']       = $this->hookName;
+      $GLOBALS[WPDKMenu::GLOBAL_MENU][$global_key]['menu_title'] = $menu_title;
     }
   }